+++ spotless-changelog-lib/src/main/java/com/diffplug/spotless/changelog/GitCfg.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2021 DiffPlug
+ * Copyright (C) 2019-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 package com.diffplug.spotless.changelog;
-
 
 import java.io.File;
 import java.io.IOException;
@@ -31,6 +30,8 @@ public class GitCfg {
 	public String commitMessage = "Published release/" + COMMIT_MESSAGE_VERSION;
 	/** Message used in tag, null means lightweight tag. */
 	public @NullOr String tagMessage = null;
+	/** Runs a CLI command after the push if not null. */
+	public @NullOr String runAfterPush = null;
 	public String remote = "origin";
 	public String branch = "main";
 	public String sshStrictHostKeyChecking = "yes";